requirements.txt
fastapi
uvicorn
twilio
python-decouple
sqlalchemy
psycopg2-binary
python-multipart
pyngrok
Flask-SQLAlchemy
flask
gTTS
pydub==0.25.1
SpeechRecognition==3.10.0
requests==2.31.0
python-dotenv
openai
boto3
langchain
datetime
pytz
gunicorn
pytest
pytest-cov
black
flake8
isort